CLN: move tools.hashing to util.hashing
