+++ asv_bench/benchmarks/algorithms.py
@@ -1,12 +1,16 @@
+from importlib import import_module
+
 import numpy as np
+
 import pandas as pd
 from pandas.util import testing as tm
 
-try:
-    from pandas.tools.hashing import hash_pandas_object
-except ImportError:
-    pass
-
+for imp in ['pandas.util.hashing', 'pandas.tools.hashing']:
+    try:
+        hashing = import_module(imp)
+        break
+    except:
+        pass
 
 class Algorithms(object):
     goal_time = 0.2
@@ -108,13 +112,13 @@ class Hashing(object):
         self.df.iloc[10:20] = np.nan
 
     def time_frame(self):
-        hash_pandas_object(self.df)
+        hashing.hash_pandas_object(self.df)
 
     def time_series_int(self):
-        hash_pandas_object(self.df.E)
+        hashing.hash_pandas_object(self.df.E)
 
     def time_series_string(self):
-        hash_pandas_object(self.df.B)
+        hashing.hash_pandas_object(self.df.B)
 
     def time_series_categorical(self):
-        hash_pandas_object(self.df.C)
+        hashing.hash_pandas_object(self.df.C)
